issue #2040 - close input streams in finally blocks
The main one was in ResourcePayload.transferTo because, without this, we
never free the non-heap memory being held by the GZIPInputStream.

+++ operation/fhir-operation-bulkdata/src/main/java/com/ibm/fhir/operation/bulkdata/config/ConfigurationAdapter.java
@@ -119,7 +119,7 @@ public interface ConfigurationAdapter {
      *
      * @return
      */
-    long getCoreCosMultiPartMinSize();
+    int getCoreCosMultiPartMinSize();
 
     /**
      * The maximum size of resources in a COS part.
